refactor(server): restructure server
Pull out functions, to give better grouping to describe the server
setup.
